backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.2.4
openai==1.54.5
httpx==0.27.0
python-dotenv==1.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
